+++ core-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao core bundle change log
 
+### DEV
+
+ * Do not log known exceptions with a pretty error screen (see #1139).
+
 ### 4.5.0-beta1 (2017-11-06)
 
  * Use ausi/slug-generator to auto-generate aliases (see #1016).